+++ src/main/java/com/cleeng/api/BatchRequest.java
@@ -1,0 +1,10 @@
+package com.cleeng.api;
+
+import org.jsonrpc.JSONRPCMessage;
+
+import java.util.List;
+
+public class BatchRequest {
+
+    public List<JSONRPCMessage> requests;
+}
