+++ src/main/java/com/cleeng/api/domain/async/AsyncCreateVodOfferRequest.java
@@ -1,5 +1,6 @@
-package com.cleeng.api;
+package com.cleeng.api.domain.async;
 
+import com.cleeng.api.AsyncRequestCallback;
 import com.cleeng.api.domain.VodOfferData;
 
 public class AsyncCreateVodOfferRequest extends AsyncRequest {
