+++ src/main/java/com/cleeng/api/domain/async/AsyncGenerateCheckoutUrlRequest.java
@@ -1,5 +1,6 @@
-package com.cleeng.api;
+package com.cleeng.api.domain.async;
 
+import com.cleeng.api.AsyncRequestCallback;
 import com.cleeng.api.domain.FlowDescription;
 
 public class AsyncGenerateCheckoutUrlRequest extends AsyncRequest {
