+++ src/main/java/com/cleeng/api/domain/async/AsyncGetAccessStatusRequest.java
@@ -1,4 +1,6 @@
-package com.cleeng.api;
+package com.cleeng.api.domain.async;
+
+import com.cleeng.api.AsyncRequestCallback;
 
 public class AsyncGetAccessStatusRequest extends AsyncRequest {
